+++ PlaneCoordinates.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.1
 Name: PlaneCoordinates
-Version: 0.1.0
+Version: 0.1.1
 Summary: Create 2D Coordinate Object types and perform some operations
 Home-page: https://github.com/rontech22/PlaneCoordinates
 Author: Ronald Rodelo
