+++ PlaneCoordinates.egg-info/SOURCES.txt
@@ -1,6 +1,7 @@
 README.md
 setup.py
 PlaneCoordinates/__init__.py
+PlaneCoordinates/module.py
 PlaneCoordinates.egg-info/PKG-INFO
 PlaneCoordinates.egg-info/SOURCES.txt
 PlaneCoordinates.egg-info/dependency_links.txt
