+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="PlaneCoordinates",                           # Nombre del paquete
-    version="0.1.0",                                # Versión inicial
+    version="0.1.1",                                # Versión inicial
     packages=find_packages(),                       # Paquetes a incluir
     description="Create 2D Coordinate Object types and perform some operations",  # Breve descripción
     author="Ronald Rodelo",                         # Tu nombre
